psffontop: Fix warnings at readpsffont

+++ src/utf8.h
@@ -1,7 +1,9 @@
 #ifndef _UTF8_H
 #define _UTF8_H
 
-extern unsigned int from_utf8(unsigned char **inptr, int cnt, int *err);
+#include <stddef.h>
+
+extern unsigned int from_utf8(unsigned char **inptr, ptrdiff_t cnt, int *err);
 
 #define UTF8_BAD (-1)
 #define UTF8_SHORT (-2)
